Use pytest plugin for --validate
This is needed so that it works both inside and outside the build dir.
https://stackoverflow.com/questions/52447767/using-pytest-addoptions-in-a-non-root-conftest-py
