+++ gsd/conftest.py
@@ -1,0 +1,3 @@
+"""Global pytest options."""
+
+pytest_plugins = ("gsd.pytest_plugin_validate",)
